refactor: move types into dedicated file
